Avoid .rpmnew when the file hasn't changed in package (rhbz#194246)
[Based on rpm.org e64bf5b9 by Panu Matilainen.]

The current behavior of %config(noreplace) creates a .rpmnewfile iff the type
of the current file has been changed wrto what was originally installed.

The patch changes this behavior so when old and new (in db and in package) is
identical -> not changed, the function returns FA_SKIP -> it won't clobber
anything, it simply skips installation of the file from the package.
This patch handles also the opposite case when old and new packages contain
%config symlink and we have regular file on disk.

Patch from Tomas Mraz.