find-requires: absolute smbolic links should not be passed to file(1)
Again, this is to improve 'buildreq -bi' support.  See, file(1) is
so full of shit that, to determine whether a link is broken, it has
to read the link and stat its destination (instead of bare stat
on the link).  This hardly makes any sense, but enables buildreq
to calculate self-dependency when the package is installed.