spec: use shuf instead of sort -R to prepare profile data
In sort -R output, identical lines adhere to each other.  Manpage says
that -R sorts by random hash of keys, which probably means that, a random
hash function, when applied to the same keys, makes the same hash value.
What we need instead is a random permutation of the input lines, though.