build/reqprov.c (compare_deps, compare_sense_flags): fix bugs in comparisons of releases
Before this change, thanks to rpmEVRcmp(), compare_deps() used to treat
EVRs with common EV as equal when one had a release and another hand't.

There were some reasons to use this approach, e.g. to decide that
	Requires: %name = %version
is weaker than
	Requires: %name = %version-%release
it had firstly to consider them as comparable deps.  But, as an
unfortunate consequence of that approach, it was deciding that such deps
are equal, so that addReqProv() was unable to replace former with the
latter.