Fix recent rpmdb handles cleanup-on signal regression
Commit 4c6e51e2c0e3deeb052ae3c47115b6d10cb0d696 seems perfectly
sensible but managed to break rpmdb handles cleanup on signals,
ie the sole reason this mess of a code exists to begin with.

The reason that change did not work for rpm is it expects the surroundings
being sane when that's not the case: rpmdbCheckTerminate() removes
the entries from the handle lists before calling the destructor
functions, which would do the very same thing if you only let them. Not
mucking with the lists from left and right simultaneously makes things
work, what a shock.

(cherry picked from commit 9b9e998c9afe33a5e2552ec03d33e13a22215394)