lib/rpmds.c: make rpmdsCompareEVR() handle DistTag
In a requirement, some components of E:V-R:D may be unspecified; here
are the sensible possibilities:

V
E:V-R
E:V-R:D

Remember that the DistTag represents the ID of a particular build. (V
represents a particular upstream version. E:V-R represents a
particular source package release.)

To satisfy a requirement, only the specified components must be
checked.

So, if the requirement doesn't specify a DistTag, then we don't have
to check the DistTag to satisfy it.

If the requirement does specify a DistTag, then if it is an "equals"
requirement, then only the same DistTag can satisfy it. (I.e., we want
that particular build.)

A "less" (or "greater") requirement of a DistTag basically makes no
sense, because the build IDs are not ordered. So, such a requirement
cannot be satisfied, if it has come to checking the DistTag, i.e. the
EVRs have been equal. (It cannot be satisfied by a package with an
equal EVR and any DistTag value, but can be satisfied by a package
with EVR which is strictly less or greater.)

What does the last part mean for "Requires" and "Conflicts"?

Requires: N>E:V-R:D would have the same effect as Requires: N>E:V-R

Conflicts: N>E:V-R:D would have the same effect as Conflicts: N>E:V-R

i.e., a conflict would not disallow packages with the specified E:V-R,
but a different DistTag. (We can't do any better, unless there is a
"not-equals" type of requirement in RPM.)

Commit message author: Ivan Zakharyaschev <imz@altlinux.org>